Create individual database benchmarking scripts
